Removed all warnings reported by VC++ 8

+++ wl_text.cpp
@@ -277,7 +277,7 @@ void HandleCommand (void)
                 //
                 // adjust this line if needed
                 //
-                if (px < leftmargin[rowon])
+                if (px < (int) leftmargin[rowon])
                         px = leftmargin[rowon];
                 break;
         }
@@ -380,7 +380,7 @@ void HandleWord (void)
         //
         VW_MeasurePropString (word,&wwidth,&wheight);
 
-        while (px+wwidth > rightmargin[rowon])
+        while (px+wwidth > (int) rightmargin[rowon])
         {
                 NewLine ();
                 if (layoutdone)
